chore: migrate marker scripts to typescript
