+++ vitest.config.js
@@ -14,7 +14,7 @@ export default defineConfig({
 		coverage: {
 			provider: "v8",
 			all: true,
-			include: ["scripts/add_marker.js"],
+			include: ["scripts/add_marker.ts"],
 			reporter: ["text", "json-summary"],
 			thresholds: {
 				branches: 100,
